Move sha256xor and scrypt chacha20poly1305 encryptor into cipher/encrypt package
