+++ src/cipher/encrypt/scrypt_chacha20poly1305_test.go
@@ -1,4 +1,4 @@
-package scryptChacha20poly1305
+package encrypt
 
 import (
 	"encoding/base64"
